requirements.txt
selenium-wire==5.1.0
selenium==4.15.2
beautifulsoup4==4.12.2
2captcha-python==1.2.0
webdriver-manager==4.0.1
xlsxwriter==3.1.9
requests==2.31.0
python-dotenv==1.0.0